oggstream: streamline tag extraction and prevent some leaks

+++ ext/ogg/gstoggstream.h
@@ -81,9 +81,9 @@ struct _GstOggStream
   int vorbis_mode_sizes[256];
   int last_size;
   int version;
-  long bitrate_upper;
-  long bitrate_nominal;
-  long bitrate_lower;
+  gint bitrate_upper;
+  gint bitrate_nominal;
+  gint bitrate_lower;
   GstTagList *taglist;
   /* theora stuff */
   gboolean theora_has_zero_keyoffset;
